+++ .github/workflows/rust-ci.yml
@@ -11,7 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        extension: ["src/core"]
+        extension: ["src/native"]
     steps:
       - uses: actions/checkout@v4
         with:
