+++ ddtrace/internal/native/__init__.py
@@ -1,0 +1,1 @@
+from ._native import DDSketch  # noqa: F401
