+++ ddtrace/internal/native/_native.pyi
