+++ ddtrace/internal/processor/stats.py
@@ -8,7 +8,7 @@ from ddtrace import config
 from ddtrace._trace.processor import SpanProcessor
 from ddtrace._trace.span import _is_top_level
 from ddtrace.internal import compat
-from ddtrace.internal.core import DDSketch
+from ddtrace.internal.native import DDSketch
 from ddtrace.internal.utils.retry import fibonacci_backoff_with_jitter
 
 from ...constants import SPAN_MEASURED_KEY
