+++ src/native/lib.rs
@@ -3,7 +3,7 @@ mod ddsketch;
 use pyo3::prelude::*;
 
 #[pymodule]
-fn _core(m: &Bound<'_, PyModule>) -> PyResult<()> {
+fn _native(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<ddsketch::DDSketchPy>()?;
     Ok(())
 }
