+++ tests/suitespec.yml
@@ -62,6 +62,7 @@ components:
     - ddtrace/internal/logger.py
     - ddtrace/internal/metrics.py
     - ddtrace/internal/module.py
+    - ddtrace/internal/native/*
     - ddtrace/internal/packages.py
     - ddtrace/internal/third-party.tar.gz
     - ddtrace/internal/periodic.py
@@ -76,7 +77,7 @@ components:
     - ddtrace/py.typed
     - ddtrace/version.py
     - ddtrace/settings/config.py
-    - src/core/*
+    - src/native/*
   datastreams:
     - ddtrace/internal/datastreams/*
     - ddtrace/data_streams.py
